requirements-web.txt
# Browser build only. Desktop-only analysis deps stay in requirements.txt.
pygbag==0.9.3
pygame-ce>=2.5.7,<3
